requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
PyYAML==5.1.2
